Tests/LibWeb/Text/expected/indexeddb-queued-delete-after-open.txt

open.upgradeneeded
open.success
conn.versionchange
conn.close() called
delete.success
DONE
